perf: support max no. possible with native data types in c

+++ fibonacci/fibonacci.c
@@ -1,4 +1,4 @@
-long fibonacci(int index){
+unsigned long long fibonacci(int index){
   if(index == 0){
     return 0;
   }
@@ -7,8 +7,8 @@ long fibonacci(int index){
     return 1;
   }
   
-  long arr[] = {0, 1};
-  long result = 0;
+  unsigned long long arr[] = {0, 1};
+  unsigned long long result = 0;
 
   for(int i=2; i<=index; i++){
     result = arr[0] + arr[1];
